+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.0.10
-- Initial basic support for linked templates
+## 0.1.0
+- Support for linked & nested templates! See readme for limitations
+- Support for the new vscode-azurearmtools extension language server and 'ARM Template' language type
+- displayName tag if present will be used in place of the resource name
 
 - ## 0.0.9
 - Many fixes to parameter & variable resolution